Moved 'ActiveSpanTracer' delegation classes to new 'activespan' subpackage 'tracer'.
No longer automatically wrap any GlobalTracer with an ActiveSpanTracer.
Separating these concerns makes the previous decision to do so feel rather random.
