+++ src/main/java/io/opentracing/contrib/activespan/tracer/ActiveSpanBuilder.java
@@ -1,4 +1,4 @@
-package io.opentracing.contrib.global;
+package io.opentracing.contrib.activespan.tracer;
 
 import io.opentracing.NoopSpanBuilder;
 import io.opentracing.Span;
